Refactor test setup to centralize initialization and enhance test control
- Created a unified `TestBase` for centralized dependency initialization, streamlining setup for all tests
- Introduced specialized test base classes (e.g., `UserVerificationTestBase`) to handle specific setup and cleanup
- Added a factory method for `UserVerificationHelper`, enabling flexible instantiation with custom configurations
- Reduced redundant setup in individual tests, ensuring tests run under consistent conditions and improving maintainability
- All tests verified to work with the new layout, providing better control and cleaner code organization
