+++ Backend.Tests/Helpers/UserServicesTestBase.cs
@@ -1,27 +1,11 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Threading.Tasks;
-using Backend.Services;
+﻿using System.Threading.Tasks;
 using Xunit;
 
-public abstract class UserServicesTestBase : IAsyncLifetime
+public abstract class UserServicesTestBase : TestBase, IAsyncLifetime
 {
-    protected readonly IServiceProvider ServiceProvider;
-    protected readonly UserServices UserServices;
-
-    protected UserServicesTestBase()
-    {
-        var serviceCollection = new ServiceCollection();
-        var startup = new StartupTest();
-        startup.ConfigureServices(serviceCollection);
-
-        ServiceProvider = serviceCollection.BuildServiceProvider();
-        UserServices = ServiceProvider.GetService<UserServices>() ?? throw new InvalidOperationException("UserServices not registered.");
-    }
-
     public async Task InitializeAsync()
     {
-        await ClearDatabaseAsync(); // Clear any pre-existing data
+        await ClearDatabaseAsync(); // Prepare a clean state
     }
 
     public async Task DisposeAsync()
@@ -31,7 +15,7 @@ public abstract class UserServicesTestBase : IAsyncLifetime
 
     private async Task ClearDatabaseAsync()
     {
-        // Delete test users and any other data set by tests
-        await UserServices.DeleteUserByEmailAsync("test@example.com"); 
+        // Custom logic to clear user data in the database
+        await UserServices.DeleteUserByEmailAsync("test@example.com");
     }
 }
